Fully automated update of Conda-Forge feedstock config.

+++ docker/sire-conda-devel/update_feedstock.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
 
 # Get the GitHub token and email.
-GITHUB_TOKEN=$1
-GITHUB_EMAIL=$2
+GITHUB_TOKEN=${GITHUB_TOKEN:-$1}
+GITHUB_EMAIL=${GITHUB_EMAIL:-$2}
 
 # Set the Conda Forge feedstock directory.
 CONDA_DIR=staged-recipes
@@ -16,37 +16,59 @@ fi
 RECIPE=$CONDA_DIR/recipes/sire/meta.yaml
 TEMPLATE=$CONDA_DIR/recipes/sire/template.yaml
 
-# List of python dependencies.
-DEPS=(boost gsl netcdf4 openmm pyqt tbb tbb-dev)
+# MacOS.
+if [ "$(uname)" = "Darwin" ]; then
+    ARCHIVE=$HOME/sire_conda_latest_osx.tar.bz2
+    SHA256=SHA256_OSX
 
-# Where the Conda environment is stored.
-CONDA_ENV=.conda_env
+# Linux.
+elif [ "$(uname)" = "Linux" ]; then
+    ARCHIVE=$HOME/sire_conda_latest_linux.tar.bz2
+    SHA256=SHA256_LINUX
 
-# Get the Sire version.
-SIRE_VER=$(git --git-dir=$HOME/Sire/.git describe --tags)
+    # List of python dependencies.
+    DEPS=(boost gsl netcdf4 openmm pyqt tbb tbb-dev)
 
-# Store the conda environment.
-$HOME/sire.app/bin/conda env export -n base > $CONDA_ENV
+    # Where the Conda environment is stored.
+    CONDA_ENV=.conda_env
 
-# Clone the feedstock repository.
-git clone --single-branch --branch devel https://github.com/michellab/staged-recipes.git > /dev/null 2>&1
+    # Get the Sire version.
+    SIRE_VER=$(git --git-dir=$HOME/Sire/.git describe --tags)
 
-# Overwite the recipe with the template file.
-cp $TEMPLATE $RECIPE
+    # Store the conda environment.
+    $HOME/sire.app/bin/conda env export -n base > $CONDA_ENV
 
-# Loop over all dependences and replace with the version installed
-# within the Conda enviroment.
-for dep in ${DEPS[@]}; do
-    ver=$(grep "\- $dep=" $CONDA_ENV | awk -F "=" '{print $2}')
-    sed -i.bak -e "s/$dep/$dep $ver/" -- $RECIPE && rm -- $RECIPE.bak
-    echo $dep $ver
-done
+    # Clone the feedstock repository.
+    git clone --single-branch --branch devel https://github.com/michellab/staged-recipes.git > /dev/null 2>&1
 
-# Update the Sire version number.
-sed -i.bak -e "s/VERSION/$SIRE_VER/" -- $RECIPE && rm -- $RECIPE.bak
+    # Overwite the recipe with the template file.
+    cp $TEMPLATE $RECIPE
 
+    # Loop over all dependences and replace with the version installed
+    # within the Conda enviroment.
+    for dep in ${DEPS[@]}; do
+        ver=$(grep "\- $dep=" $CONDA_ENV | awk -F "=" '{print $2}')
+        sed -i.bak -e "s/$dep/$dep $ver/" -- $RECIPE && rm -- $RECIPE.bak
+    done
+
+    # Update the Sire version number.
+    sed -i.bak -e "s/VERSION/$SIRE_VER/" -- $RECIPE && rm -- $RECIPE.bak
+
+    # Remove the Conda environment file.
+    rm -f .conda_env
+
+# Unsupported OS.
+else
+    echo "Unsupported operating system: '$uname'"
+    exit -1
+fi
+
+# Update the sha256 checksum.
+CHECKSUM=$(openssl sha256 $ARCHIVE | awk '{print $2}')
+sed -i.bak -e "s/$SHA256/$CHECKSUM/" -- $RECIPE && rm -- $RECIPE.bak
+
+# Remove the recipe backup file.
 rm -f $RECIPE.bak
-rm -f .conda_env
 
 # Change to Conda package directory and update git config.
 cd $CONDA_DIR
